Baseline for stargate exporter
