+++ bin/g
@@ -28,7 +28,7 @@ JSON_DATA=$(cat <<EOF
 {
         "model": "gpt-3.5-turbo",
         "messages": ${PROMPT},
-        "temperature": 0.5,
+        "temperature": 0.25
 }
 EOF
          )
