+++ lib/constants.py
@@ -56,3 +56,5 @@ ETHEREUM = "ethereum"
 BITCOIN = "bitcoin"
 ARBITRUM_ONE = "arbitrum-one"
 ARBITRUM_NOVA = "arbitrum-nova"
+POLYGON = "polygon"
+BSC = "BSC"
