+++ lib/stargate/constants.py
@@ -4,8 +4,8 @@ Provide constants to be used in stargate exporter
 
 ############## ADDRESSES ##############
 
-Router = "0x8731d54E9D02c286767d56ac03e8037C07e01e98".lower()
-ETH_Router = "0x150f94B44927F078737562f0fcF3C95c01Cc2376".lower()
+ROUTER = "0x8731d54E9D02c286767d56ac03e8037C07e01e98".lower()
+ETH_ROUTER = "0x150f94B44927F078737562f0fcF3C95c01Cc2376".lower()
 
 ############## SIGNATURES ##############
 # swap sig
@@ -15,3 +15,4 @@ SEND_CREDITS_SIG = "0x6939f93e3f21cf1362eb17155b740277de5687dae9a83a85909fd71da9
 
 # Swap: user_address -> Router
 # SendCredits: Router -> user_address
+# Withdraw not found
