+++ lib/stargate/stargate.py
@@ -25,4 +25,11 @@ class StargateExporter(Exporter):
 
     def run(self):
         """The main function to run the Stargate exporter"""
+        self.start_logging()
+        records = self.read_records()
+        if not records:
+            logging.info("No records found for %s", self.name)
+            return
+        processed_records = process(project_name=self.name, records=records)
+        self.insert_records(processed_records)
